remove javax.persistence dependency from keycloak-services
