+++ services/src/main/java/org/keycloak/credential/AAGUIDConverter.java
@@ -14,19 +14,15 @@
  * limitations under the License.
  */
 
-package org.keycloak.models.jpa.converter;
+package org.keycloak.credential;
 
 import com.webauthn4j.data.attestation.authenticator.AAGUID;
 
-import javax.persistence.AttributeConverter;
-
-public class AAGUIDConverter implements AttributeConverter<AAGUID, byte[]> {
-    @Override
+public class AAGUIDConverter {
     public byte[] convertToDatabaseColumn(AAGUID aaguid) {
         return aaguid.getBytes();
     }
 
-    @Override
     public AAGUID convertToEntityAttribute(byte[] bytes) {
         return new AAGUID(bytes);
     }
