+++ services/src/main/java/org/keycloak/credential/AttestationStatementConverter.java
@@ -14,25 +14,21 @@
  * limitations under the License.
  */
 
-package org.keycloak.models.jpa.converter;
+package org.keycloak.credential;
 
 import com.webauthn4j.converter.util.CborConverter;
-import com.webauthn4j.data.attestation.authenticator.CredentialPublicKey;
+import com.webauthn4j.data.attestation.statement.AttestationStatement;
 import com.webauthn4j.util.Base64UrlUtil;
 
-import javax.persistence.AttributeConverter;
+public class AttestationStatementConverter {
 
-public class CredentialPublicKeyConverter implements AttributeConverter<CredentialPublicKey, String> {
+    private CborConverter converter = new CborConverter();
 
-    private CborConverter converter = new CborConverter(); //TODO: Inject by CDI to make it singleton
-
-    @Override
-    public String convertToDatabaseColumn(CredentialPublicKey credentialPublicKey) {
-        return Base64UrlUtil.encodeToString(converter.writeValueAsBytes(credentialPublicKey));
+    public String convertToDatabaseColumn(AttestationStatement attribute) {
+        return Base64UrlUtil.encodeToString(converter.writeValueAsBytes(attribute));
     }
 
-    @Override
-    public CredentialPublicKey convertToEntityAttribute(String s) {
-        return converter.readValue(Base64UrlUtil.decode(s), CredentialPublicKey.class);
+    public AttestationStatement convertToEntityAttribute(String dbData) {
+        return converter.readValue(Base64UrlUtil.decode(dbData), AttestationStatement.class);
     }
 }
